+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/model/PackagesPrefs.java
@@ -31,7 +31,8 @@ public class PackagesPrefs extends JavaScriptObject
                                   boolean viewDirAfterCheckFailure,
                                   boolean hideObjectFiles,
                                   boolean useDevtools,
-                                  boolean useSecureDownload) /*-{
+                                  boolean useSecureDownload,
+                                  boolean useNewlineInMakefiles) /*-{
       var prefs = new Object();
       prefs.cran_mirror = cranMirror;
       prefs.use_internet2 = useInternet2;
@@ -41,6 +42,7 @@ public class PackagesPrefs extends JavaScriptObject
       prefs.hide_object_files = hideObjectFiles;
       prefs.use_devtools = useDevtools;
       prefs.use_secure_download = useSecureDownload;
+      prefs.use_newline_in_makefiles = useNewlineInMakefiles;
       return prefs ;
    }-*/;
 
@@ -74,6 +76,10 @@ public class PackagesPrefs extends JavaScriptObject
    
    public native final boolean getUseSecureDownload() /*-{
       return this.use_secure_download;
+   }-*/;
+   
+   public native final boolean getUseNewlineInMakefiles() /*-{
+      return this.use_newline_in_makefiles;
    }-*/;